+++ src/main/java/obro1961/chatpatches/mixin/codec/ExtraCodecsMixin.java
@@ -1,32 +1,36 @@
 package obro1961.chatpatches.mixin.codec;
 
-import com.llamalad7.mixinextras.expression.Definition;
-import com.llamalad7.mixinextras.expression.Expression;
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
-import com.mojang.serialization.DataResult;
+import com.llamalad7.mixinextras.sugar.Local;
 import net.minecraft.util.ExtraCodecs;
 import obro1961.chatpatches.ChatLog;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
-
-import java.util.function.Function;
 
 // we can't comment out the entire file otherwise older versions will throw errors bc the mixin won't exist
 @Mixin(ExtraCodecs.class)
 public abstract class ExtraCodecsMixin {
 	//? if >=1.21.5 {
+	@Unique private static final String LAMBDA_METHOD_NAME = "method_66032";
+
 	/**
-	 * Allows all chat strings to be serialized while restoring chat logs. Fixes
+	 * Allows all chat strings to be serialized despite any section signs present.
+	 * Unfortunately, a sole check to {@link ChatLog#isCodecSafe()} will still break
+	 * context menu clicks on said messages, so a more comprehensive check is
+	 * required. Fixes
 	 * <a href="https://github.com/mrbuilder1961/ChatPatches/issues/246">#246</a>,
 	 * which seems to only affect 1.21.5+.
+	 *
+	 * @implNote Targets the synthetic method corresponding to the lambda in {@link
+	 * ExtraCodecs#CHAT_STRING}'s initializer.
 	 */
-	@Definition(id = "CHAT_STRING", field = "Lnet/minecraft/util/ExtraCodecs;CHAT_STRING:Lcom/mojang/serialization/Codec;")
-	@Definition(id = "STRING", field = "Lcom/mojang/serialization/Codec;STRING:Lcom/mojang/serialization/codecs/PrimitiveCodec;")
-	@Definition(id = "validate", method = "Lcom/mojang/serialization/codecs/PrimitiveCodec;validate(Ljava/util/function/Function;)Lcom/mojang/serialization/Codec;")
-	@Expression("CHAT_STRING = STRING.validate(@(?))")
-	@ModifyExpressionValue(method = "<clinit>", at = @At("MIXINEXTRAS:EXPRESSION"))
-	private static Function<String, DataResult<String>> allowEverythingWhileRestoring(Function<String, DataResult<String>> checker) {
-		return ChatLog.isCodecSafe().get() ? checker : DataResult::success;
+	@ModifyExpressionValue(
+		method = LAMBDA_METHOD_NAME + "(Ljava/lang/String;)Lcom/mojang/serialization/DataResult;",
+		at = @At(value = "INVOKE", target = "Lnet/minecraft/util/StringUtil;isAllowedChatCharacter(C)Z")
+	)
+	private static boolean allowSectionSigns(boolean isValidChatCharacter, @Local char c) {
+		return isValidChatCharacter || c == '§';
 	}
 	//? }
 }