+++ changelog.md
@@ -4,8 +4,12 @@
 - ACTUALLY fixed 'Disallowed character' toast and log spam ([#246](https://www.github.com/mrbuilder1961/ChatPatches/issues/246))
 - Now requires Fabric Loader `0.17.0`+ (doesn't currently apply, but will also require NeoForge `21.1.18`+)
 - Override chat height now explicitly changes the **focused** chat height, and the docs + config suggest this ([#254](https://www.github.com/mrbuilder1961/ChatPatches/issues/254))
+- Chat log now instead dumps the current chat to `chatlog_dump_<time>.json` instead of to the log. This can be renamed to `chatlog.json` to restore the log 
+  if needed (FIXME: doesn't work yet)
+- Removed log spam from messages being formatted incorrectly ("`DUPE_INDEX` is out of bounds")
 
 ## Chat Patches `8.0-alpha.3` for Minecraft 1.20.1-1.21.8 on Fabric, Quilt
+- Fixed the main crash from [#250](https://www.github.com/mrbuilder1961/ChatPatches/issues/250)
 - Fixed the search bar accidentally including invisible formatting codes in the search results
 - Fixed the search bar not initially working with certain settings
 - Fixed the context menu using incorrect messages when searching